Get rid of PrivateUdpServer (unnecessary layer of abstraction)
