--- .checkstyle
@@ -1,2 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<fileset-config file-format-version="1.0.0"/>
