+++ emulinker/src/main/java/org/emulinker/kaillera/relay/V086Relay.kt
@@ -17,6 +17,7 @@ import org.emulinker.util.EmuUtil.dumpBuffer
 
 private val logger = FluentLogger.forEnclosingClass()
 
+// TODO(nue): What is this..
 class V086Relay
     @AssistedInject
     internal constructor(
